some fix for TPA and horizon mode!

+++ AeroQuad/AeroQuad.h
@@ -79,7 +79,7 @@ float userRateRollP = 0.0;
 float userRateRollD = 0.0;
 float userRatePitchP = 0.0;
 float userRatePitchD = 0.0;
-int throttlePIDAdjustmentFactor = 0.0;
+int throttlePIDAdjustmentFactor = 0;
 
 // main loop time variable
 unsigned long previousTime = 0;
@@ -265,7 +265,7 @@ typedef struct {
   
   // Gyro calibration
   float ROTATION_SPEED_FACTOR_ARD;
-  float THROTTLE_PID_ADJUSTMENT_ADR;
+  long  THROTTLE_PID_ADJUSTMENT_ADR;
   // Accel Calibration
   float XAXIS_ACCEL_BIAS_ADR;
   float XAXIS_ACCEL_SCALE_FACTOR_ADR;
